requirements/devrequirements.txt
django-extensions==1.3.3
Werkzeug==0.9.4
flexmock==0.9.7
nose==1.3.0
pinocchio==0.3.1
should-dsl==2.1.2
ipdb==0.8
coverage==3.7.1
nose-exclude==0.2.0
coveralls==0.4.1
django-nose==1.2
nosier==1.1
